pyproject.toml
[tool.poetry]
name = "chatbot"
version = "0.1.0"
description = "Chatbot backend service"
authors = ["Sidhanta Pradhan <sidd07dev@gmail.com>"]
readme = "README.md"
packages = [{ include = "chatbot", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.12,<4.0.0"
fastapi = "^0.128.6"
sqlalchemy = "^2.0.46"
pydantic = "^2.12.5"
python-dotenv = "^1.2.1"
uvicorn = "^0.40.0"
psycopg2-binary = "^2.9.11"
openai = "^2.20.0"
openai-agents = "^0.8.3"
redis = "^7.1.1"
langchain-text-splitters = "^1.1.0"

[tool.poetry.scripts]
dev = "src.chatbot.main:start"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"